third_party/WebKit/LayoutTests/sensor/idl-LinearAccelerationSensor.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
test(function() {
  // Test that LinearAccelerationSensor interface exists
  assert_true('LinearAccelerationSensor' in window);
}, 'Test that the LinearAccelerationSensor interface is present');
</script>
